Added testsuite
