+++ cmd/dummy/dummy.go
@@ -26,7 +26,7 @@ import (
 	_ "go.uber.org/cadence/activity"
 	_ "go.uber.org/cadence/client"
 	_ "go.uber.org/cadence/encoded"
-	_ "go.uber.org/cadence/testing"
+	_ "go.uber.org/cadence/testsuite"
 	_ "go.uber.org/cadence/worker"
 	_ "go.uber.org/cadence/workflow"
 )
